USB now a little more reliable, hashing is still broken
What happens when USB is reset now affects the state of condominer.c;
namely, we flush all requests from cgminer when the USB is reset from
cgminer.

About Endianness: the previous commit has incorrect info in the comment.
PIC32MX CPU is Little Endian, while the SHA256 routine SHOULD be Big
Endian. It's not certain how the cgminer driver driver-klondike.c sends
midstate and merkle data to the firmware: is the data Little Endian or
Big Endian?

+++ framework/system/common/sys_module.h
@@ -99,7 +99,7 @@ typedef unsigned short int SYS_MODULE_INDEX;
     type.
 */
 
-typedef uintptr_t SYS_MODULE_OBJ;
+typedef int SYS_MODULE_OBJ;
 
 
 // *****************************************************************************
